code/requirements.txt
snowflake-snowpark-python
langchain
javalang